+++ common/src/main/java/net/mehvahdjukaar/slotify/mixins/AbstractContainerScreenMixin.java
@@ -2,6 +2,7 @@ package net.mehvahdjukaar.slotify.mixins;
 
 import com.llamalad7.mixinextras.injector.WrapWithCondition;
 import com.llamalad7.mixinextras.sugar.Local;
+import net.mehvahdjukaar.slotify.MenuModifier;
 import net.mehvahdjukaar.slotify.MenuModifierManager;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.screens.Screen;
@@ -15,11 +16,29 @@ import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+import static net.mehvahdjukaar.slotify.MenuModifierManager.MENU_MODIFIERS;
 
 @Mixin(AbstractContainerScreen.class)
 public abstract class AbstractContainerScreenMixin<T extends AbstractContainerMenu> extends Screen implements MenuAccess<T> {
 
-    @Shadow @Final protected T menu;
+    @Shadow
+    @Final
+    protected T menu;
+
+    @Shadow
+    protected int titleLabelY;
+
+    @Shadow
+    protected int titleLabelX;
+
+    @Shadow
+    protected int inventoryLabelY;
+
+    @Shadow
+    protected int inventoryLabelX;
 
     protected AbstractContainerScreenMixin(Component component) {
         super(component);
@@ -30,13 +49,32 @@ public abstract class AbstractContainerScreenMixin<T extends AbstractContainerMe
             value = "INVOKE"
     ))
     public boolean slotifyColor(GuiGraphics guiGraphics, int x, int y, int blitOffset,
-                             @Local Slot slot){
+                                @Local Slot slot) {
         MenuType<?> type;
-        try{
+        try {
             type = this.menu.getType();
-        }catch (Exception e){
+        } catch (Exception e) {
             type = null;
         }
         return MenuModifierManager.maybeChangeColor(type, slot, guiGraphics, x, y, blitOffset);
+    }
+
+    // this could be done with events...
+    @Inject(method = "init", at = @At("TAIL"))
+    public void modifyLabels(CallbackInfo ci) {
+        MenuType<?> type;
+        try {
+            type = this.menu.getType();
+        } catch (Exception e) {
+            type = null;
+        }
+        //ugly, but we need to access these fields
+        MenuModifier m = MENU_MODIFIERS.get(type);
+        if (m != null) {
+            this.titleLabelX += m.titleX();
+            this.titleLabelY += m.titleY();
+            this.inventoryLabelX += m.labelX();
+            this.inventoryLabelY += m.labelY();
+        }
     }
 }